Added an actual win and draw window
I made it so that when someone wins or a draw happens a dialog (its like a small panel) pops up and says who won or for a draw it says the reason for the draw. You can then close that window or restart the game, which will also automatically close the window.